clean up on the code base
