+++ app/Http/Controllers/HomeController.php
@@ -196,216 +196,14 @@ foreach ($get as $key => $value) {
     }
 
 
-    public function publish(Request $request,$username) {
-        $title = isset($request->title) ? $request->title : '';
-        $content = $request->postVal;
-        $tags = $request->tags;
-        $action = $request->action;
-
-
-          $initial_images = array_filter($request->all(), function ($key) {
-            return preg_match('/^img-\w*$/', $key);
-        }, ARRAY_FILTER_USE_KEY);
-
-        $images = [];
-        foreach ($initial_images as $key => $value) {
-            $newKey = preg_replace('/_/', '.', $key);
-            $images[$newKey] = $value;
-            // Log::debug($value);
-
-        }
-        $post = new \Lucid\Core\Document($username);
-        $createPostAction = $post->createPost($title, $content, $tags, $images,$username,$action);
-
-      //  dd(  $createPost);
-        if($createPostAction == "publish" ){
-          return response()->json(["error" => false, "action"=>"publish", "message" => "Post published successfully"],200);
-        }else if ($createPostAction == "draft") {
-          return response()->json(["error" => false, "action"=>"draft", "message" => "Post published successfully"],200);
-        }else if($createPostAction == false){
-          return response()->json(["error" => true, "action"=>"publish", "message" => "Failed while processing your request, please try again"]);
-        }
-    }
+    
 
 
 
 
-    public function settings(){
-      $user = Auth::user();
-      $username = $user['username'];
-      // follower and following Count
-      $sub = new \Lucid\Core\Subscribe($username);
-      $fcount =$sub->myfollowercount();
-      $count = $sub->count();
-      //dd($fcount);
-      if (!empty($fcount)) {
-          $fcount = count($fcount);
-        }else {
-          $fcount = "";
-        }
-        if (!empty($count)) {
-          $count = count($count);
-        }else {
-          $count = "";
-        }
+    
 
-
-        //User Follower checker
-        if(Auth::user()){
-          $check = new \Lucid\Core\Subscribe(Auth::user()->username);
-          $fcheck = $check->followCheck($user->name);
-        }
-        else {
-          $fcheck = "no";
-        }
-
-      return view('settings', ['fcheck' => $fcheck,'user'=>$user,'fcount' => $fcount , 'count' => $count ]);
-
-    }
-
-    public function saveSettings(Request $request) {
-          $validator=Validator::make($request->all(),[
-            'name' => 'required',
-            'email' => ['required','email',
-             Rule::unique('users')->ignore(Auth::user()->id),
-            ],
-            'profileimage' => 'image|mimes:jpeg,png,jpg,gif|max:2048',
-            'user_id' =>'required',
-            'username'=>['required',
-             Rule::unique('users')->ignore(Auth::user()->id),
-            ]
-        ]);
-
-      if($validator->fails()){
-          return response()->json($validator->messages(), 200);
-      }
-
-      $oldname = Auth::user()->name;
-      $newname = $request->name;
-      $user_id = $request->user_id;
-      $email = $request->email;
-      $username= Str::slug($request->username, '_');
-      $bio = $request->bio;
-      $FolderName = storage_path('app/'.Auth::user()->id);
-
-      if(!is_null($request->file('profileimage')) && $request->file('profileimage') !== ""){
-
-        $filenamewithextension = $request->file('profileimage')->getClientOriginalName();
-
-        //get filename without extension
-        $filename = pathinfo($filenamewithextension, PATHINFO_FILENAME);
-
-        //get file extension
-        $extension = $request->file('profileimage')->getClientOriginalExtension();
-
-        $image = $request->file('profileimage');
-        $fullPath = $this->store($image, $filenamewithextension,$filename,$extension);
-
-        //  $url = Auth::user()->id."/images/";
-
-//         $path = Storage::disk('public')->put($url, $request->file('profileimage'));
-  //       $fullPath = '/storage/'.$path;
-
-         $updated =   DB::transaction(function ()
-   use ($oldname,$newname,$fullPath,$FolderName,$user_id,$email,$username,$bio) {
-
-  $updated= DB::table('users')->where('id',$user_id)
-    ->update(['name'=>$newname,'username'=>$username,'email'=>$email,'image'=>$fullPath,'short_bio'=>$bio]);
-
-DB::table('extfeeds')->where('site',$oldname)
-    ->update([
-      'site'=>$newname,
-      'site_image' => $fullPath
-    ]);
-
-return true;
-
-   });
-
-        if($updated) {
-
-          return response()->json(['success'=>"Your changes has been saved successfully",'img_path'=>$fullPath,'renamedUserContentFolderName'=>$username], 200);
-        }
-      } else {
-        $fullPath = Auth::user()->image;
-
-        $updated =   DB::transaction(function ()
-   use ($oldname,$newname,$fullPath,$FolderName,$user_id,$email,$username,$bio) {
-
-     DB::table('users')->where('id',$user_id)
-                ->update(['name'=>$newname,'username'=>$username,'email'=>$email,'short_bio'=>$bio,'image' => $fullPath]);
-
-    DB::table('ext_rsses')->where('title',$oldname)
-                ->update([
-                  'title'=>$newname,
-                  'url'=> $FolderName."/rss/rss.xml",
-                  'link'=> $FolderName."/rss/rss.xml",
-                  'image' => $fullPath
-                ]);
-return true;
-
-});
-                                      if($updated){
-                                        return response()->json(['success'=>"Your changes has been saved successfully",'renamedUserContentFolderName'=>$username], 200);
-                                      }
-      }
-
-    }
-
-    public function store($image,$filenamewithextension,$filename,$extension)
-{
-
-
-        //filename to store
-        $filenametostore = $filename.'_'.time().'.'.$extension;
-
-        //small thumbnail name
-        $smallthumbnail = $filename.'_small_'.time().'.'.$extension;
-
-        //medium thumbnail name
-        $mediumthumbnail = $filename.'_medium_'.time().'.'.$extension;
-
-        //large thumbnail name
-        $largethumbnail = $filename.'_large_'.time().'.'.$extension;
-
-        //Upload File
-        $image->storeAs('public/'.Auth::user()->id.'/images/', $filenametostore);
-        $image->storeAs('public/'.Auth::user()->id.'/images/thumbnail', $smallthumbnail);
-        $image->storeAs('public/'.Auth::user()->id.'/images/thumbnail', $mediumthumbnail);
-        $image->storeAs('public/'.Auth::user()->id.'/images/thumbnail', $largethumbnail);
-
-        //create small thumbnail
-        $smallthumbnailpath = public_path('storage/'.Auth::user()->id.'/images/thumbnail/'.$smallthumbnail);
-        $this->createThumbnail($smallthumbnailpath, 150, 93);
-
-        //create medium thumbnail
-        $mediumthumbnailpath = public_path('storage/'.Auth::user()->id.'/images/thumbnail/'.$mediumthumbnail);
-        $this->createThumbnail($mediumthumbnailpath, 300, 185);
-
-        //create large thumbnail
-        $largethumbnailpath = public_path('storage/'.Auth::user()->id.'/images/thumbnail/'.$largethumbnail);
-        $this->createThumbnail($largethumbnailpath, 550, 340);
-
-        $imagePath ='/storage/'.Auth::user()->id.'/images/thumbnail/'.$smallthumbnail;
-        return $imagePath;
-
-}
-        /**
-        * Create a thumbnail of specified size
-        *
-        * @param string $path path of thumbnail
-        * @param int $width
-        * @param int $height
-        */
-        public function createThumbnail($path, $width, $height)
-        {
-        $img = Image::make($path)->resize($width, $height, function ($constraint) {
-        $constraint->aspectRatio();
-        });
-        $img->save($path);
-
-        //$img = Image::make($path)->resize($width, $height)->save($path);
-        }
+    
 
 
 
@@ -451,29 +249,7 @@ return true;
 
     }
 
-    public function deletePost(Request $request, $username){
-       $post = DB::table('posts')->where('id',$request->post_id)->first();
-       $parsedown  = new Parsedown();
-       $postContent = $parsedown->text($post->content);
-       preg_match_all('/<img[^>]+src="((\/|\w|-)+\.[a-z]+)"[^>]*\>/i', $postContent, $matches);
-       foreach($matches[1] as $found_img) {
-         $image_name_array = explode('/',$found_img);
-         $img_name = end($image_name_array);
-         $imagePath = storage_path('app/public/'.$post->user_id.'/images/'.$img_name);
-         if(file_exists($imagePath)) {
-           unlink($imagePath);
-         }
-       }
-       DB::table('notifications')->where('post_id',$post->id)->delete();
-        DB::table('extfeeds')->where('title',$post->title)->delete();
-      $deletePost = DB::table('posts')->where('id',$post->id)->delete();
-      if($deletePost) {
-        return response()->json(['success'=>"Post Successfully Deleted"],200);
-      }else{
-        return response()->json(['error'=>"Something not right"],500);
-      }
-
-    }
+    
 
     // debuging tools
 public function checkpost($value)
@@ -579,44 +355,8 @@ print_r($updateFeeds);
     }
 
 
-    public function editPost(Request $request, $username) {
+    
 
-        $title = isset($request->title) ? $request->title : '';
-        $content = $request->postVal;
-        $tags = $request->tags;
-        $post_id = $request->post_id;
-        $action = $request->action;
-
-
-          $initial_images = array_filter($request->all(), function ($key) {
-            return preg_match('/^img-\w*$/', $key);
-        }, ARRAY_FILTER_USE_KEY);
-
-        $images = [];
-        foreach ($initial_images as $key => $value) {
-            $newKey = preg_replace('/_/', '.', $key);
-            $images[$newKey] = $value;
-        }
-        $post = new \Lucid\Core\Document($username);
-        $updatePost = $post->saveUpdatedPost($title, $content, $tags, $images,$username,$post_id,$action);
-
-        if($updatePost){
-          return response()->json(["error" => false, "action"=>"update", "message" => "Post Updated successfully"],200);
-        }else{
-          return response()->json(["error" => true, "action"=>"error", "message" => "Fail while publishing, please try again"]);
-        }
-    }
-
-    public function updatePostStatus($username,$post_id,$action){
-      $update = DB::table('posts')->where('id',$post_id)->update([
-        'action'=>$action,
-      ]);
-
-      if($update) {
-        return response()->json(['success'=>true],200);
-      }else{
-        return response()->json(['success'=>false],200);
-      }
-    }
+    
 
 }
