+++ routes/web.php
@@ -26,8 +26,8 @@ Route::get('register', function () {
 });
 
 Route::prefix('explore')->group(function (){
-    Route::get('/','pageController@explorePage');
-    Route::get('/interest/{interest}','pageController@interest')->name('interest');
+    Route::get('/','ExploreController@explorePage');
+    Route::get('/interest/{interest}','ExploreController@interest')->name('interest');
 });
 Route::get('loader', function () {
     return view('preloader');
@@ -51,8 +51,8 @@ Route::get('microblog','HomeController@microblog');
 Route::post('save-post','HomeController@savePost');
 Route::post('save-subscription','pageController@saveSubscriptionEmail');
 
-Route::get('/category/{category}','pageController@postCategories');
-Route::get('/filter/{method}','pageController@filterPost');
+Route::get('/category/{category}','ExploreController@postCategories');
+Route::get('/filter/{method}','ExploreController@filterPost');
 
 Route::get('/sitemap_users.xml','pageController@sitemapUsers');
 Route::get('/sitemap_feeds.xml','pageController@sitemapFeeds');
@@ -68,15 +68,15 @@ Route::prefix('{username}')->group(function () {
 
     //get Request
     Route::get('/contact', 'pageController@contact');
-    Route::get('/post/{postTitle}','pageController@singlePostPage')->name('post');
-    Route::get('/post-data/{id}','pageController@getPostData');
-    Route::get('/','pageController@homePage');
-    Route::get('/home','pageController@homePage');
+    Route::get('/post/{postTitle}','PostController@singlePostPage')->name('post');
+    Route::get('/post-data/{id}','PostController@getPostData');
+    Route::get('/','PostController@homePage');
+    Route::get('/home','PostController@homePage');
     Route::get('/thoughts','pageController@thoughts');
     Route::get('/logout', "Auth\LoginController@logout");
-    Route::get('/posts','pageController@posts');
+    Route::get('/posts','PostController@posts')->middleware('auth');
     Route::get('/subscribe','HomeController@subscribe');
-    Route::get('/settings', 'HomeController@settings');
+    Route::get('/settings', 'UserAccountSettingsController@settings');
     Route::get('/followers','pageController@followers')->name("followers");
     Route::get('/following','pageController@following')->name("following");
     Route::get('/comments/{post_id}','pageController@comments')->name('comment');
@@ -85,7 +85,7 @@ Route::prefix('{username}')->group(function () {
     Route::get('/like','ReactionsController@like');
     Route::get('/love','ReactionsController@love');
     Route::get('/feeds','pageController@Feeds');
-    Route::get('/update-post-status/{post_id}/{action}','HomeController@updatePostStatus');
+    Route::get('/update-post-status/{post_id}/{action}','PostController@updatePostStatus')->middleware('auth');
 
 
 
@@ -95,14 +95,14 @@ Route::prefix('{username}')->group(function () {
     Route::post('/addrss','ExtRssController@addRss');
     Route::post('/unfollow','ExtRssController@unfollow');
     Route::post('/extrss','ExtRssController@addExtRss');
-    Route::post('/publish','HomeController@publish');
+    Route::post('/publish','PostController@publish')->middleware('auth');
     Route::post('/send-mail','SendEmailController@sendEmail');
-    Route::post('/save_settings','HomeController@saveSettings');
+    Route::post('/save_settings','UserAccountSettingsController@saveSettings');
     Route::post('/update-contact-details','HomeController@updateContactDetails');
-    Route::post('/delete-post','HomeController@deletePost')->name('deletePost');
+    Route::post('/delete-post','PostController@deletePost')->name('deletePost')->middleware('auth');
     Route::post('/save-comment','HomeController@saveComment')->name('save-comment');
     Route::post('/notif','pageController@notification');
-    Route::post('/edit-post','HomeController@editPost');
+    Route::post('/edit-post','PostController@editPost')->middleware('auth');
